Consertando bug em Arquivo.

+++ Framework/src/main/java/com/lab/data/banco/Banco.sql
@@ -60,7 +60,7 @@ create table Arquivo (
     codigoArquivo int auto_increment not null,
     nome varchar(255) not null,
     tipo varchar(255),
-    tamanho int not null,
+    tamanho bigint not null,
     codigoProjeto int not null,
     codigoPratica int,
     codigoDados int not null,
